Create a new input.rs module for handling user interactions
